Attempted to clean up cable.py by changing the calculate_segments method
I was unsuccessful in trying to get my cluster algorithm to work. This was mainly because the data structure did not allow for connections between houses. I was not successful in trying to change the data structure, due to its complexity. I did manage to clean up some code and increase its readability.